# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Category API",
        "description": "API endpoints for managing categories",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/deals": {
            "get": {
                "tags": [
                    "Deals"
                ],
                "summary": "Get a list of deals",
                "operationId": "f7e426f6a75edf51dea85d09a8ed1c2d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Deals"
                ],
                "summary": "Create a new deal",
                "operationId": "7742b8aca5f3ace96ef1f79e2585b998",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "buyer_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    },
                                    "deal_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deal created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/deals/{id}": {
            "get": {
                "tags": [
                    "Deals"
                ],
                "summary": "Get a specific deal",
                "operationId": "d82ca02ead88b497b75314b20583bcb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Deal not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Deals"
                ],
                "summary": "Update a specific deal",
                "operationId": "5d2b9fbb26cc77a959cc377b3b82b261",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "buyer_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    },
                                    "deal_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deal updated successfully"
                    },
                    "404": {
                        "description": "Deal not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Deals"
                ],
                "summary": "Delete a specific deal",
                "operationId": "2186eb3e73a874926faca4151e39ab40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deal deleted successfully"
                    },
                    "404": {
                        "description": "Deal not found"
                    }
                }
            }
        },
        "/api/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get a list of likes",
                "operationId": "297b9c3ea341e74585a2ae193ef6dcee",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Create a new like",
                "operationId": "511c8161eb44f7b3479a914692d975ee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userlike_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Like created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/likes/{id}": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get a specific like",
                "operationId": "cc6d000adc2ca5c218da958d798e920a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Like not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Likes"
                ],
                "summary": "Update a specific like",
                "operationId": "4fca184960c72abcfe28c03f60e7e12b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userlike_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like updated successfully"
                    },
                    "404": {
                        "description": "Like not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Delete a specific like",
                "operationId": "e02fff678edcb4bc50ab9c57d7e4531d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Like deleted successfully"
                    },
                    "404": {
                        "description": "Like not found"
                    }
                }
            }
        },
        "/api/users/{userId}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get likes created by a specific user",
                "operationId": "ddebfb13a8679a2dacbe18aa6f05926e",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
...
}